rename module bapsflib._hdf.maps.controls.contype -> bapsflib._hdf.maps.controls.types
