+++ bapsflib/_hdf/maps/controls/n5700ps.py
@@ -20,7 +20,7 @@ import warnings
 from bapsflib.utils.errors import HDFMappingError
 from warnings import warn
 
-from .contype import ConType
+from .types import ConType
 from .templates import HDFMapControlCLTemplate
