+++ bapsflib/_hdf/maps/controls/sixk.py
@@ -21,7 +21,7 @@ import re
 from bapsflib.utils.errors import HDFMappingError
 from warnings import warn
 
-from .contype import ConType
+from .types import ConType
 from .templates import HDFMapControlTemplate
