+++ bapsflib/_hdf/maps/controls/waveform.py
@@ -21,7 +21,7 @@ import warnings
 from bapsflib.utils.errors import HDFMappingError
 from warnings import warn
 
-from .contype import ConType
+from .types import ConType
 from .templates import HDFMapControlCLTemplate
