+++ bapsflib/_hdf/utils/helpers.py
@@ -388,7 +388,7 @@ def condition_controls(hdf_file: File,
         #. Input **controls** should be
            :code:`Union[str, Iterable[Union[str, Tuple[str, Any]]]]`
         #. There can only be one control for each
-           :class:`~bapsflib._hdf.maps.controls.contype.ConType`.
+           :class:`~bapsflib._hdf.maps.controls.types.ConType`.
         #. If a control has multiple configurations, then one must be
            specified.
         #. If a control has ONLY ONE configuration, then that will be
