--- docs/api_static/bapsflib._hdf.maps.controls.contype.rst
@@ -1,9 +1,0 @@
-bapsflib\.\_hdf\.maps\.controls\.contype
-========================================
-
-.. py:currentmodule:: bapsflib._hdf.maps.controls.contype
-
-.. automodapi:: bapsflib._hdf.maps.controls.contype
-    :no-heading:
-    :include-all-objects:
-    :headings: "-^"
