+++ docs/api_static/bapsflib._hdf.maps.controls.types.rst
@@ -1,0 +1,9 @@
+bapsflib\.\_hdf\.maps\.controls\.types
+======================================
+
+.. py:currentmodule:: bapsflib._hdf.maps.controls.types
+
+.. automodapi:: bapsflib._hdf.maps.controls.types
+    :no-heading:
+    :include-all-objects:
+    :headings: "-^"
